Refactor: Clean up and restructure repository
This commit addresses two main user requests:
1.  Cleaning up the repository by removing redundant and unnecessary files.
2.  Restructuring the application to better separate concerns.

Changes:
-   Deleted temporary and generated files (`okx_data.log`, `pylint_report.txt`, `BTC-USDT_historical.json`).
-   Removed redundant data directory `data/BTC-USDT/medium/`.
-   Replaced legacy `SupportResistanceAnalysis` with `NewSupportResistanceAnalysis` and updated all call sites.
-   Standardized the currency configuration by removing the `DEFAULT_SYMBOLS` list and using `WATCHLIST` as the single source of truth.
-   Standardized the symbol format in the configuration to `XXX-USDT`.
-   Restructured the `src` directory by creating `src/indicators` and renaming `src/data` to `src/data_retrieval`.
-   Moved indicator-related modules to the new `src/indicators` directory.
-   Updated all import paths to reflect the new directory structure.
-   Updated all tests to reflect the changes and ensure all tests pass.
