+++ src/data_retrieval/base_fetcher.py
